requirements.txt
requests>=2.31.0
urllib3>=2.0.7
certifi>=2023.7.22
charset-normalizer>=3.3.2
idna>=3.4
pyyaml>=6.0.1 